UPDATE: rework generate barcode methods - Ws.Labels.Service

+++ Src/App/Domain/Ws.Domain.Models/ValueTypes/BarcodeItem.cs
@@ -4,7 +4,7 @@ public class BarcodeItem
 {
     public string Property { get; set; } = string.Empty;
     public string FormatStr { get; set; } = string.Empty;
-    public int Length { get; set; }
+    public ushort Length { get; set; }
     public bool IsConst { get; set; }
 
     public override bool Equals(object? obj)
